 requirements.txt
feedparser
tencentcloud-sdk-python-tmt
requests
pyyaml